requirement.txt
confluent_kafka
pyserial